rename internals.predictive to internals.nmc
