+++ tests/robust/test_ops.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 from typing_extensions import ParamSpec
 
-from chirho.robust.internals.predictive import PredictiveFunctional, PredictiveModel
+from chirho.robust.internals.nmc import PredictiveFunctional, PredictiveModel
 from chirho.robust.ops import influence_fn
 
 from .robust_fixtures import SimpleGuide, SimpleModel
